refactor: component relationship with store
