+++ package.json
@@ -86,7 +86,7 @@
     "mdi-react": "7.5.0",
     "object-hash": "2.2.0",
     "preact": "10.5.13",
-    "prismjs": "1.23.0",
+    "prismjs": "1.24.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-hot-toast": "2.0.0",
@@ -201,6 +201,7 @@
   "resolutions": {
     "**/**/dns-packet": "5.2.2",
     "**/**/browserslist": "4.16.5",
+    "**/**/prismjs": "1.24.0",
     "@stacks/ui": "7.9.0",
     "@stacks/connect-ui": "5.1.5",
     "@stacks/ui-core": "7.3.0",
