+++ src/common/asset-types.ts
