+++ src/pages/add-network/add-network.tsx
@@ -5,7 +5,7 @@ import { Formik } from 'formik';
 import { currentNetworkKeyState, networksState } from '@store/networks';
 import { PopupContainer } from '@components/popup/container';
 import { useDoChangeScreen } from '@common/hooks/use-do-change-screen';
-import { ScreenPaths } from '@store/common/types';
+import { ScreenPaths } from '@common/types';
 import { isValidUrl } from '@common/validation/validate-url';
 import { ChainID, fetchPrivate } from '@stacks/transactions';
 import { ErrorLabel } from '@components/error-label';
