+++ src/pages/home/home.tsx
@@ -3,8 +3,8 @@ import { Stack } from '@stacks/ui';
 import { PopupContainer } from '@components/popup/container';
 import { Header } from '@components/header';
 import { BalancesAndActivity } from '@components/popup/balances-and-activity';
-import { UserAccount } from '@components/home/components/user-area';
-import { HomeActions } from '@components/home/components/actions';
+import { UserAccount } from '@pages/home/components/user-area';
+import { HomeActions } from '@pages/home/components/actions';
 
 const PageTop = () => (
   <>
