+++ src/pages/set-password.tsx
@@ -5,7 +5,7 @@ import { Form, Formik } from 'formik';
 import * as yup from 'yup';
 import { PopupContainer } from '@components/popup/container';
 import { useDoChangeScreen } from '@common/hooks/use-do-change-screen';
-import { ScreenPaths } from '@store/common/types';
+import { ScreenPaths } from '@common/types';
 import { useWallet } from '@common/hooks/use-wallet';
 
 import { useOnboardingState } from '@common/hooks/auth/use-onboarding-state';
