+++ src/pages/signed-out/signed-out-view.tsx
@@ -4,7 +4,7 @@ import { Box, Button } from '@stacks/ui';
 import { PopupContainer } from '@components/popup/container';
 import { Text } from '@components/typography';
 
-import { ScreenPaths } from '@store/common/types';
+import { ScreenPaths } from '@common/types';
 import { useDoChangeScreen } from '@common/hooks/use-do-change-screen';
 import { Header } from '@components/header';
